[SPMD] Allow Override.replacement to have a global view
It's easier for users to implement one Override that takes care of
all target submodules of different types, instead of specifying one
mapping pair for each FQN/type. For example, when calculating
sharding for sparse layers, the decision needs to be make globally.
In this, case it's helpful to allow user Override to get access to
all submodules and make replacement decisions accordingly.

ghstack-source-id: e71a0b28a1b0c20d85a24488268b6529c9de40a9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/101427