[MTPG] Enable for some tests in test_fsdp_misc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102043

Enables MTPG for some FSDP tests in this file. Tests that need the
backward pass and warning logging are left as follow up work.

Backward pass issue: it seems that the registered MTPG is not recognized during
collectives in the backward pass, which makes sense since the collective is in
a different thread.

Warning issue: We have a couple tests that regex check on warnings, but in the
multithreaded scenario these warnings are somehow not logged.
ghstack-source-id: 190352125

Differential Revision: [D43209769](https://our.internmc.facebook.com/intern/diff/D43209769/)